Working topic model
Suppose you have a DDT domain, say, "Domain_Name". Associated to that
domain, you have a collection of documents. After you launch DDT
at localhost:8084 (and possibly refresh the list of documents in the
main pane), you can go to

    http://localhost:8084/topic_model?ddt_domain=domain_name

And obtain some tokens from the first document. This is a test to see
that the topic model is working.