Topic visualization works now
Make sure you have no ad blockers of any sort active in your browser. I
had Privacy Badger installed and it was stopping the visualization from
loading!

To use pyLDAvis to visualize your model, go to

  http://localhost:8084/topicsvis?ddt_domain=ddt_domain_name

Where `ddt_domain_name` is the name of the domain your using.

TODO 1: use session information to load this. The topicsvis method that
runs the model currently takes the ddt_domain_name argument directly. We
need to put session info there (there are other Cherrrypy-exposed
methods in vis/server.py that will provide a blueprint of how to do
that). To obtain session info for testing/debugging open the developer
tools (I used Chrome dev tools this time) and go to the Network log.
Load the main page

  http://localhost:8084

And filter to find info on the `getPagesSummary` call. Look in there in
the Headers tab, Form Data -> session. There you will find the session
info (mainly the domainID) that you can use for testing and debugging
purposes.

TODO 2: it would be nicer to not save the data from the visualization
anywhere, maybe just the model. The pyLDAvis package provides means for
doing that. Topik has its own notion of a project.  Ideally the saved
model should be an abstraction over the specific algos (so at the level
of topik, not of pyLDAvis). Need to look into that.