Bypasses possible timeouts due to topik projects
Before I was having timeouts when trying to save the models
(TopikProject objects) to Elasticsearch. I moved to the current
formulation, and started getting core dumps.

I reinstalled pyldavis and topik, and, for some reason, things work now,
in the current formulation.

I will go back and check if this reinstallation fixed the problem when
saving the model to Elasticsearch.

PLSA still does not work.