Extracted modeling and viz from exposed method
We were afraid that having modeling visualization inside an exposed
method was causing Cherrypy to serialize objects that it shouldn't be
serializing. We suspected this due to a call to

site-packages/cherrypy/lib/encoding.py

To test that hypothesis, I took out the modeling and visualization
generation out of the exposed method (Pages.topicsvis) and put into a
non-exposed method (Pages.topicsvis_aux).

That did not work. Therefore, we can revert to the previous commit if
desired.